# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorCode": {
        "enum": [
          "IDENTIFIER_REQUIRED",
          "IDENTIFIER_CHARACTERS_INVALID",
          "IDENTIFIER_LENGTH_UNSUPPORTED",
          "IDENTIFIER_CHECK_DIGIT_INVALID",
          "PACKAGE_MATCH_SOURCE_UNAVAILABLE",
          "REFERENCE_IMAGE_URL_INVALID",
          "REFERENCE_IMAGE_NOT_FOUND",
          "REFERENCE_IMAGE_SOURCE_UNAVAILABLE",
          "RATE_LIMIT_EXCEEDED",
          "SEARCH_QUERY_REQUIRED",
          "SEARCH_QUERY_INVALID",
          "SEARCH_RESULTS_NOT_FOUND"
        ],
        "title": "ErrorCode",
        "type": "string"
      },
      "ErrorDetail": {
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorDetail",
        "type": "object"
      },
      "ErrorEnvelope": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
...
      "ExternalDatasetVersionResponse": {
        "properties": {
          "activated_at": {
            "format": "date-time",
            "title": "Activated At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "retrieved_at": {
            "format": "date-time",
            "title": "Retrieved At",
            "type": "string"
          },
          "sha256": {
            "title": "Sha256",
            "type": "string"
          },
          "source_url": {
            "title": "Source Url",
            "type": "string"
          }
        },
        "required": [
          "id",
          "source_url",
          "retrieved_at",
          "activated_at",
          "sha256"
        ],
        "title": "ExternalDatasetVersionResponse",
        "type": "object"
      },
...
      "JsonValue": {
        "anyOf": [
          {
            "type": "boolean"
          },
          {
            "type": "integer"
          },
          {
            "type": "number"
          },
          {
            "type": "string"
          },
          {
            "items": {
              "$ref": "#/components/schemas/JsonValue"
            },
            "type": "array"
          },
          {
            "additionalProperties": {
              "$ref": "#/components/schemas/JsonValue"
            },
            "type": "object"
          },
          {
            "type": "null"
          }
        ]
      },
...
      "PackageMatchEvidenceResponse": {
        "properties": {
          "dataset_version_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dataset Version Id"
          },
          "field": {
            "title": "Field",
            "type": "string"
          },
          "language": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Language"
          },
          "observed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Observed At"
          },
          "retrieved_at": {
            "format": "date-time",
            "title": "Retrieved At",
            "type": "string"
          },
          "source_field": {
            "title": "Source Field",
            "type": "string"
          },
          "source_name": {
            "title": "Source Name",
            "type": "string"
          },
          "source_revision": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Revision"
          },
          "source_url": {
            "title": "Source Url",
            "type": "string"
          },
          "value": {
            "$ref": "#/components/schemas/JsonValue"
          }
        },
        "required": [
          "field",
          "value",
          "source_field",
          "source_name",
          "source_url",
          "language",
          "observed_at",
          "retrieved_at"
        ],
        "title": "PackageMatchEvidenceResponse",
        "type": "object"
      },
...
      "PackageSearchAdditiveResponse": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "code",
          "name"
        ],
        "title": "PackageSearchAdditiveResponse",
        "type": "object"
      },
      "PackageSearchNameResponse": {
        "properties": {
          "language": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Language"
          },
          "source_field": {
            "title": "Source Field",
            "type": "string"
          },
          "value": {
            "title": "Value",
            "type": "string"
          }
        },
        "required": [
          "value",
          "language",
          "source_field"
        ],
        "title": "PackageSearchNameResponse",
        "type": "object"
      },
      "PackageSearchResponse": {
        "properties": {
          "dataset_version": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ExternalDatasetVersionResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "has_more": {
            "title": "Has More",
            "type": "boolean"
          },
          "page": {
            "title": "Page",
            "type": "integer"
          },
          "page_size": {
            "title": "Page Size",
            "type": "integer"
          },
          "query": {
            "title": "Query",
            "type": "string"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/PackageSearchResultResponse"
            },
            "title": "Results",
            "type": "array"
          },
          "source": {
            "title": "Source",
            "type": "string"
          }
        },
        "required": [
          "query",
          "page",
          "page_size",
          "has_more",
          "source",
          "dataset_version"
        ],
        "title": "PackageSearchResponse",
        "type": "object"
      },
      "PackageSearchResultResponse": {
        "properties": {
          "additives": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/PackageSearchAdditiveResponse"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Additives"
          },
          "allergens": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Allergens"
          },
          "barcode": {
            "title": "Barcode",
            "type": "string"
          },
          "brand": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Brand"
          },
          "dataset_version_id": {
            "title": "Dataset Version Id",
            "type": "string"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/PackageMatchEvidenceResponse"
            },
            "title": "Evidence",
            "type": "array"
          },
          "image_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Url"
          },
          "ingredients": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ingredients"
          },
          "made_in": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Made In"
          },
          "matched_fields": {
            "items": {
              "type": "string"
            },
            "title": "Matched Fields",
            "type": "array"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "other_names": {
            "items": {
              "$ref": "#/components/schemas/PackageSearchNameResponse"
            },
            "title": "Other Names",
            "type": "array"
          },
          "quantity": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quantity"
          },
          "source_record_id": {
            "title": "Source Record Id",
            "type": "string"
          },
          "source_revision": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Revision"
          },
          "source_updated_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Updated At"
          },
          "source_url": {
            "title": "Source Url",
            "type": "string"
          }
        },
        "required": [
          "barcode",
          "name",
          "source_record_id",
          "source_url",
          "source_updated_at",
          "dataset_version_id"
        ],
        "title": "PackageSearchResultResponse",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "LifeGoods API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/package-matches/search": {
      "get": {
        "description": "Searches the active local Open Food Facts Dataset Version by product name, brand, or explicit manufacturing country. Results are external Evidence and do not prove product origin or identity.",
        "operationId": "searchPackageMatches",
        "parameters": [
          {
            "description": "Product name, brand, or manufacturing country to search for.",
            "in": "query",
            "name": "q",
            "required": true,
            "schema": {
              "description": "Product name, brand, or manufacturing country to search for.",
              "maxLength": 100,
              "minLength": 2,
              "title": "Q",
              "type": "string"
            }
          },
          {
            "description": "Result page number.",
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "description": "Result page number.",
              "maximum": 100,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "description": "Results per page.",
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 20,
              "description": "Results per page.",
              "maximum": 50,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PackageSearchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "No Package Match candidates matched the search query."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Search Package Match candidates",
        "tags": [
          "Package Matches"
        ]
      }
    }
  }
}